@@ -1,1 +1,1 @@
-module.exports = require('./lib/geo');
+(function(a,b){if('function'==typeof define&&define.amd)define(['module','exports','lodash/isFunction'],b);else if('undefined'!=typeof exports)b(module,exports,require('lodash/isFunction'));else{var c={exports:{}};b(c,c.exports,a.isFunction),a.geo=c.exports}})(this,function(a,b,c){'use strict';function d(a,b){var c;return{create:function(d,e){c=setTimeout(function(){a.clearWatch(d),(0,g.default)(e)&&e('Timed out')},b)},remove:function(){clearTimeout(c)}}}function e(a,b){return function(c){var d,e;e=c.coords.latitude,d=c.coords.longitude,(0,g.default)(a)&&a(e,d,c),(0,g.default)(b)&&b()}}function f(a,b){return function(c){(0,g.default)(a)&&a(c),(0,g.default)(b)&&b()}}Object.defineProperty(b,'__esModule',{value:!0}),b.default=function(a){var b,c,g,i,j,k;if(g=h(),!!g)return a&&(k=d(g,a),i=k.remove),function(a,d){j=e(a,i),b=f(d,i),c=g.watchPosition(j,b),k&&k.create(c,d)}};var g=function(a){return a&&a.__esModule?a:{default:a}}(c),h=function(){return navigator.geolocation};a.exports=b['default']});
